+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.4.0]
+- Notifications4Plugin library
+- Fix recommendation multiple modal
+
 ## [1.3.5]
 - Fixes